+++ core/user_2fa_otp_web_register.js
@@ -5,6 +5,7 @@
 const Config = require('./config.js').get;
 const getServer = require('./listening_server.js').getServer;
 const webServerPackageName = require('./servers/content/web.js').moduleInfo.packageName;
+const { WellKnownLocations } = require('./servers/content/web');
 const {
     createToken,
     deleteToken,
@@ -76,7 +77,8 @@ module.exports = class User2FA_OTPWebRegister {
                 (token, textTemplate, htmlTemplate, callback) => {
                     const webServer = getWebServer();
                     const registerUrl = webServer.instance.buildUrl(
-                        `/_internal/enable_2fa_otp?token=${token}&otpType=${otpType}`
+                        WellKnownLocations.Internal +
+                            `/2fa/enable_2fa_otp?token=${token}&otpType=${otpType}`
                     );
 
                     const replaceTokens = s => {
@@ -170,7 +172,9 @@ module.exports = class User2FA_OTPWebRegister {
                     return User2FA_OTPWebRegister.accessDenied(webServer, resp);
                 }
 
-                const postUrl = webServer.instance.buildUrl('/_internal/enable_2fa_otp');
+                const postUrl = webServer.instance.buildUrl(
+                    WellKnownLocations.Internal + '/2fa/enable_2fa_otp'
+                );
                 const config = Config();
                 return webServer.instance.routeTemplateFilePage(
                     _.get(config, 'users.twoFactorAuth.otp.registerPageTemplate'),
@@ -296,12 +300,12 @@ ${backupCodes}
         [
             {
                 method: 'GET',
-                path: /^\/_internal\/enable_2fa_otp\?token=[a-f0-9]+&otpType=[a-zA-Z0-9_]+$/,
+                path: /^\/_enig\/2fa\/enable_2fa_otp\?token=[a-f0-9]+&otpType=[a-zA-Z0-9_]+$/,
                 handler: User2FA_OTPWebRegister.routeRegisterGet,
             },
             {
                 method: 'POST',
-                path: /^\/_internal\/enable_2fa_otp$/,
+                path: /^\/_enig\/2fa\/enable_2fa_otp$/,
                 handler: User2FA_OTPWebRegister.routeRegisterPost,
             },
         ].forEach(r => {