+++ WHATSNEW.md
@@ -4,7 +4,7 @@ This document attempts to track **major** changes and additions in ENiGMA½. For
 ## 0.0.14-beta
 * The [Web Server](/docs/_docs/servers/contentservers/web-server.md) has made some possibly breaking changes:
     * `/static/` prefixes are no longer required. This was a ugly hack.
-    * Some internal routes such as those used for password resets live within `/_internal/`.
+    * Some internal routes such as those used for password resets live within `/_enig/`.
     * Routes for the file base now default to `/_f/` prefixed instead of just `/f/`. If `/f/` is in your `config.hjson` you are encouraged to update it!
     * Finally, the system will search for `index.html` and `index.htm` in that order, if another suitable route cannot be established.